Works.
OK, so the trick is to create the Watch App as a C# project (which is meaningless, since it's just a storyboard).

Then, you have to edit the .storyboard file manually, adding actions and outlets. Regrettably, XCode will delete these every time you open the storyboard, so that's a problem!